Fix Java wrapping parallel build issue
Currently the VTK Java classes are compiled using generated “driver”
classes that reference every class in a particular module. This
"driver" class is compiled and because of the references javac
compiles the associated classes. In a parallel build multiple
invocation of javac are used to compile the “driver” classes, these
interact in some way, probably a race condition to build dependencies.
The use of the “driver” classes is unnecessary. All Java classes can
be compiled by a single invocation of javac passing in *.java, javac
will then compile all classes and dependencies. As there is a single
invocation the parallel build issue it resolved. This also simplifies
the CMakeLists.txt. There is room to refactor further by removing
some of the build targets and unnecessary dependencies.

Change-Id: I7beb8ffef0972611d5a78150e1c52e340993d740